Make grid size to integer for GEMM example

+++ examples/gemm/sgemm.py
@@ -39,7 +39,7 @@ def sgemm(A, B,
     code = read_code(sgemm_file, params=config)
     kern = load_kernel('sgemm', code)
 
-    grid = (math.ceil(m / blk_m), math.ceil(n / blk_n), 1)
+    grid = (int(math.ceil(m / blk_m)), int(math.ceil(n / blk_n)), 1)
     block = (dim_x, dim_y, 1)
     args = (m, n, k, A, B, C)
     shared_mem = blk_k * (blk_m + 1) * 4 + blk_n * (blk_k + 1) * 4
